+++ com.unity.render-pipelines.high-definition/Runtime/Material/Decal/DecalSystem.cs
@@ -37,10 +37,11 @@ namespace UnityEngine.Rendering.HighDefinition
         // Caution: Enum num must match pass name for s_MaterialDecalPassNames array
         public enum MaterialDecalPass
         {
-            DBufferProjector = 0,
-            DecalProjectorForwardEmissive = 1,
-            DBufferMesh = 2,
-            DecalMeshForwardEmissive = 3,
+            ScenePickingPass = 0,
+            DBufferProjector = 1,
+            DecalProjectorForwardEmissive = 2,
+            DBufferMesh = 3,
+            DecalMeshForwardEmissive = 4,
         };
 
         public static readonly string[] s_MaterialDecalPassNames = Enum.GetNames(typeof(MaterialDecalPass));